+++ packages/react-toc/src/components/icons/CloseIcon.tsx
