+++ packages/react-toc/src/index.ts
@@ -1,10 +1,10 @@
-import "./style.css";
+import "@/styles/style.css";
 
-export { H2 } from "./H2";
-export { H3 } from "./H3";
-export { H4 } from "./H4";
-export { H5 } from "./H5";
-export { H6 } from "./H6";
-export { Toc } from "./Toc";
-export { TocProvider } from "./TocProvider";
-export { TocTopBar } from "./TocTopBar";
+export { H2 } from "@/components/headings/H2";
+export { H3 } from "@/components/headings/H3";
+export { H4 } from "@/components/headings/H4";
+export { H5 } from "@/components/headings/H5";
+export { H6 } from "@/components/headings/H6";
+export { Toc } from "@/components/Toc";
+export { TocProvider } from "@/components/TocProvider";
+export { TocTopBar } from "@/components/TocTopBar";
