+++ packages/react-toc/tsconfig.json
@@ -4,7 +4,10 @@
     "outDir": "dist",
     "rootDir": "./src",
     "declaration": true,
-    "declarationDir": "dist"
+    "declarationDir": "dist",
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   },
   "include": ["src"]
 }
